+++ src/pages/api/login/admin/sessions/[id].js
@@ -2,32 +2,37 @@ import SessionModel from "@/app/models/session";
 import database_connection from "@/app/database/mongodb";
 
 export default async function handler(req, res) {
-    const { id } = req.query;
+    const url = req.url;
+    const urlSplit = url.split("/");
+    const email = req.body.email;
+    const user = await UserModel.findOne({ email });
+    if(urlSplit[3] === user.role) {
+        const { id } = req.query;
 
-    await database_connection();
+        await database_connection();
 
-    switch (req.method) {
-        case "PUT":
-            try {
-                const { newCours: cours, newDays: days, newMonths: months, newYears: years } = req.body;
+        if (req.method === "PUT") {
+                try {
+                    const { newCours: cours, newDays: days, newMonths: months, newYears: years } = req.body;
 
-               
-                const updatedSession = await SessionModel.findByIdAndUpdate(
-                    id,
-                    { cours, days, months, years },
-                    { new: true }
-                );
+                
+                    const updatedSession = await SessionModel.findByIdAndUpdate(
+                        id,
+                        { cours, days, months, years },
+                        { new: true }
+                    );
 
-                if (!updatedSession) {
-                    return res.status(404).json({ error: "Session introuvable" });
+                    if (!updatedSession) {
+                        return res.status(404).json({ error: "Session introuvable" });
+                    }
+
+                    return res.status(200).json({ message: "La Session a été mis à jour", sessions: updatedSession });
+                } catch (error) {
+                    console.error(error);
+                    return res.status(500).json({ error: "Erreur lors de la mise à jour" });
                 }
-
-                return res.status(200).json({ message: "La Session a été mis à jour", sessions: updatedSession });
-            } catch (error) {
-                console.error(error);
-                return res.status(500).json({ error: "Erreur lors de la mise à jour" });
-            }
-            case "DELETE":
+        }
+        if (req.method === "PUT") {
             try {
                 const deletedSession = await SessionModel.findByIdAndDelete(id);
 
@@ -40,9 +45,11 @@ export default async function handler(req, res) {
                 console.error(error);
                 return res.status(500).json({ error: "Erreur lors de la suppression" });
             }
-
-        default:
-            res.setHeader("Allow", ["PUT", "DELETE"]);
-            return res.status(405).json({ error: `Méthode ${req.method} non autorisée` });
+        }
     }
+    
+    if (urlSplit[3] !== user.role) {
+        return res.status(500).json({ error: "Vous n'êtes pas un professeur, vous ne pouvez pas accedez à ces données" });
+    }
+
 }
